auth

Module user_provider

Modules

static_user_provider 🔒

watch_file_user_provider 🔒

Traits

UserProvider

Functions

authenticate_with_credential 🔒

load_credential_from_file 🔒

parse_credential_line 🔒

Parse a line of credential in the format of username=password or username:permission_mode=password

Type Aliases

UserInfoMap

Type alias for user info map Key is username, value is (password, permission_mode)