Crate auth

Modules

common 🔒

error

permission 🔒

user_info 🔒

user_provider 🔒

Structs

DefaultPermissionChecker

The default permission checker implementation. It checks the permission mode of DefaultUserInfo.

PermissionTableTarget

A user-visible table target for permission checks.

PgScramSha256Verifier

StaticUserProvider

Enums

Identity

Password

Authentication information sent by the client.

PermissionReq

PermissionResp

PermissionTableTargets

The result of resolving every table target in a permission request.

PgAuthInfo

Constants

DEFAULT_PBKDF2_SHA256_ITERATIONS

DEFAULT_PBKDF2_SHA256_SALT_LEN

MAX_PBKDF2_SHA256_ITERATIONS

MAX_PBKDF2_SHA256_SALT_LEN

PBKDF2_SHA256_HASH_LEN

PG_SCRAM_SHA256_KEY_LEN

Traits

PermissionChecker

UserInfo

UserProvider

Functions

auth_mysql

format_mysql_native_password_verifier

format_pbkdf2_sha256_password_verifier

format_pg_scram_sha256_password_verifier

mysql_native_password_hash

static_user_provider_from_option

user_provider_from_option

userinfo_by_name

construct a UserInfo impl with name use default username greptime if None is provided

Type Aliases

HashedPassword

PermissionCheckerRef

UserInfoRef

pub type alias

UserProviderRef